Português/Projeto Canino/requirements.txt
Pillow
requests
pandas
xlrd
tqdm
numpy
opencv-python
matplotlib
scikit-learn
tensorflow
tensorflow_hub